even more cleaning up
