--- apigen/www/cambio_metodo.php
@@ -1,5 +1,0 @@
-<?php
-session_start();
-$_SESSION["metodo"]=$_GET["id"];
-header("Location: editmethod.inc.php");
-?>
