+++ apigen/www/inc/bottom.inc.php
@@ -1,0 +1,20 @@
+
+                <div class="mtm pvm uiBoxWhite topborder">
+                    <div class="mbm"> </div>
+                </div>
+
+            </div>
+
+            <div class="clear">
+            </div>
+
+        </div>
+    </div>
+    <div class="footer">
+        <div class="content">
+        </div>
+    </div>
+</div>
+
+</body>
+</html>
