+++ apigen/www/newmethod.php
@@ -1,0 +1,8 @@
+<?php require_once("inc/top.inc.php"); ?>
+
+<hr/>
+<?php require_once( "inc/newmethod.inc.php" ); ?>
+<hr/>
+
+<?php require_once("inc/bottom.inc.php"); ?>
+
